pyproject.toml
[project]
name = "sdd-jps-engine"
version = "4.0.0"
description = "SDD-JPS Engine v4.0.0 — unified MCP server with 73+ tools for spec-driven development"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "fpdf2>=2.8.0",
    "markdown>=3.5",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
    "ruff>=0.5.0",
]

[project.scripts]
sdd-jps-engine = "server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
